pyproject.toml
[project]
name = "civic-slm"
version = "0.0.0"
description = "Domain-specialized fine-tune of Qwen2.5-7B for California municipal documents."
readme = "README.md"
requires-python = ">=3.11,<3.13"
authors = [{ name = "Duncan", email = "itsmeduncan@gmail.com" }]
license = { text = "Apache-2.0" }

dependencies = [
    "pydantic>=2.7",
    "typer>=0.12",
    "structlog>=24.1",
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "rich>=13.7",
    "pyyaml>=6.0.3",
]

[project.optional-dependencies]
ingest = [
    "browser-use>=0.1",
    "playwright>=1.45",
    "pypdf>=4.3",
]
synth = [
    "anthropic>=0.40",
]
train = [
    "mlx>=0.18; sys_platform == 'darwin' and platform_machine == 'arm64'",
    "mlx-lm>=0.19; sys_platform == 'darwin' and platform_machine == 'arm64'",
    "datasets>=2.20",
    "wandb>=0.17",
]
eval = [
    "sentence-transformers>=3.0",
]
dev = [
    "ruff>=0.6",
    "pyright>=1.1.380",
    "pytest>=8.3",
    "pytest-cov>=5.0",
]

[project.scripts]
civic-slm = "civic_slm.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/civic_slm"]

[dependency-groups]
dev = [
    "ruff>=0.6",
    "pyright>=1.1.380",
    "pytest>=8.3",
    "pytest-cov>=5.0",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --strict-markers"
filterwarnings = ["error"]
asyncio_mode = "auto"
